drivers: i2c_ll_stm32_v2: add support for optional timing property
In case optional timing property has been defined in DT, let's use it
instead of using runtime algorithm. The value will be set directly to
the I2C_TIMINGR register through the LL API.

This property is only valid for I2C V2 peripheral, so the timing config
structure member is only added in case of CONFIG_I2C_STM32_V2. Also the
initialization of the member is done conditionnaly, setting to 0 in case
it is not defined in DT. (0 not being a valid timing configuration for
an I2C link)

Signed-off-by: Laurent Meunier <laurent.meunier@st.com>